R2-gating: cast the PCA input to float32
Selin: "we want float32", extended to add_pca's descriptive all-cells fit as well as
to the per-fold CV fits.

SPLIT OUT DELIBERATELY, because it sequences the opposite way to the rest of the R4
work it was written alongside. This cast changes what PREPROCESSING produces --
`X_pca` moves in its last digits -- so **R2 runs after it**. Everything else on
`worktree-r4-training-changes` must NOT land before R2, since R2 exists to test the
restructured preprocessing before the training path moves under it. Bundled, they
could only land together, which is wrong in both directions.

WHY float32 AT ALL. The per-fold CV PCA reads a dense 53,513 x 5,000 counts matrix;
in float64 that is 2.14 GB resident and ~4.4 GB peak once a fold's training subset is
taken. float32 halves both. The cast is extended to this descriptive fit so that it
and the per-fold fits stay identical in kind -- preserving what Selin established on
10.08.2026 when she harmonized `ddof` to 1, that the two fits differ only in which
cells they see. Casting only the fold fits would have quietly undone that.

The precision cost is nil for PCA on log-CPM; the visible consequence is that `X_pca`
is not bit-identical to the values currently on disk, which is why this gates R2
rather than riding along with it.

The identical hunk also sits on worktree-r4-training-changes and is deliberately left
there: that branch's `fold_pca_projections` docstring states both fits are float32, so
removing it would make that branch self-inconsistent in isolation. Once this lands on
main and R4 rebases, the two are the same content and resolve without conflict.

+++ scripts/preprocessing/add_pca.py
@@ -246,7 +246,14 @@ def run(
 
     # The train-fitted keys are computed FIRST, and read src.X directly: sc.pp.scale below
     # standardizes src.X in place, so afterwards it is no longer the transformed matrix.
-    X_log = src.X.toarray() if sparse.issparse(src.X) else np.asarray(src.X)
+    # float32 (Selin, 12.08.2026). Chosen for the per-fold CV fits, where the counts matrix is
+    # 53,513 x 5,000 and float64 doubles peak memory to ~4.4 GB, and extended here so the
+    # descriptive all-cells fit and the per-fold fits stay identical in kind -- the property
+    # she established on 10.08.2026 when she harmonized `ddof` to 1, that the two differ only in
+    # which cells they see. It is a preprocessing change: `X_pca` moves in its last digits.
+    X_log = (src.X.toarray() if sparse.issparse(src.X) else np.asarray(src.X)).astype(
+        np.float32, copy=False
+    )
     for col in split_cols:
         key = train_pca_key(col)
         if key not in todo:
